backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
apscheduler==3.11.0
attrs==25.3.0
beanie==1.29.0
beautifulsoup4==4.13.3
blurry-cli==0.14.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
dpath==2.2.0
email-validator==2.2.0
faiss-cpu==1.10.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastmcp==0.4.1
ffmpeg-python==0.2.0
filetype==1.2.0
frozendict==2.4.6
frozenlist==1.5.0
future==1.0.0
greenlet==3.1.1
groq==0.20.0
h11==0.14.0
htmlmin2==0.1.13
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jinja2==3.1.6
jinja2-simple-tags==0.6.1
jinjax==0.48
jiter==0.9.0
js2py-3-13==0.74.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.47
langchain-groq==0.3.1
langchain-mcp-adapters==0.0.5
langchain-openai==0.3.9
langchain-tavily==0.1.5
langchain-text-splitters==0.3.7
langgraph==0.3.18
langgraph-checkpoint==2.0.23
langgraph-prebuilt==0.1.4
langgraph-sdk==0.1.58
langsmith==0.3.18
lazy-model==0.2.0
livereload==2.7.1
lxml==5.3.1
m3u8==6.0.0
markdown-it-py==3.0.0
markupsafe==3.0.2
marshmallow==3.26.1
mcp==1.4.1
mdurl==0.1.2
mistune==3.1.3
motor==3.7.0
multidict==6.2.0
mypy==1.15.0
mypy-extensions==1.0.0
numpy==2.2.4
openai==1.68.2
orjson==3.10.16
ormsgpack==1.9.0
packaging==24.2
pillow==11.2.1
propcache==0.3.0
pydantic==2.11.2
pydantic-core==2.33.1
pydantic-settings==2.8.1
pygments==2.19.1
pyjsparser==2.7.1
pyld==2.0.4
pymongo==4.11.3
pyotp==2.9.0
pypdf==5.4.0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
selectolax==0.3.28
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
sqlalchemy==2.0.39
sse-starlette==2.2.1
starlette==0.46.1
tenacity==9.0.0
tiktoken==0.9.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
twikit==2.3.3
typer==0.15.2
typing-extensions==4.13.1
typing-inspect==0.9.0
typing-inspection==0.4.0
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
wand==0.6.13
watchfiles==1.0.4
websockets==15.0.1
webvtt-py==0.5.1
yarl==1.18.3
zstandard==0.23.0
